backend/requirements.txt
asgiref==3.11.1
attrs==26.1.0
autobahn==25.12.2
Automat==25.4.16
cbor2==5.9.0
certifi==2024.8.30
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
charset-normalizer==3.4.0
constantly==23.10.4
coreapi==2.3.3
coreschema==0.0.4
crispy-bootstrap5==2024.10
cryptography==46.0.7
daphne==4.2.1
Django==5.1.2
django-cors-headers==4.6.0
django-crispy-forms==2.3
django-extensions==4.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==24.1.1
hyperlink==21.0.0
idna==3.10
Incremental==24.11.0
itypes==1.2.0
Jinja2==3.1.4
Markdown==3.7
MarkupSafe==3.0.2
msgpack==1.1.2
packaging==26.0
pillow==11.1.0
py-ubjson==0.16.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
PyJWT==2.9.0
pyOpenSSL==26.0.0
redis==7.4.0
requests==2.32.3
service-identity==24.2.0
sqlparse==0.5.1
Twisted==25.5.0
txaio==25.12.2
typing_extensions==4.15.0
tzdata==2024.2
ujson==5.12.0
uritemplate==4.1.1
urllib3==2.2.3
whitenoise==6.9.0
zope.interface==8.3
